added sound options for each alert type, option to not show new sub alerts for gifts, and color changer for preview

+++ database/database_utilities/layouts/deleteLayout.js
@@ -1,15 +1,25 @@
 const db = require('../../models');
 
-const deleteLayout = async (layoutId) => {
+const deleteLayout = async (layoutId, streamerDatabaseId) => {
     try {
+
+        const streamer = await db.Streamer.findOne({
+            where: {
+                id: streamerDatabaseId
+            }
+        });
+        const idInt = parseInt(layoutId);
+        if (idInt === streamer.active_slideshow_layout || streamer.active_alerts_layouts.includes(idInt)) {
+            return "Error: Layout is active"
+        }
         await db.Layout.destroy({
             where: {
                 id: layoutId
             }
         })
-        return;
+        return 'successfully deleted layout';
     } catch (error) {
-        console.error('Error deleting layout', error)
+        console.error('Error deleting layout', error);
     }
 }
